Clean up
- move master details page to its own file
- commit pubspec.lock for now, keeping an eye on it
